pyproject.toml
[project]
name = "astmgr"
version = "0.0.1"
authors = [
  { name="Steven Joseph", email="steven@contextu.al" },
]
description = "Assistant manager will help you manage."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "docopt",
    "python-box",
    "jira",
    "python-gitlab",
    "PyYAML",
    "tabulate",
    "jmespath",
]

[project.urls]
"Homepage" = "https://github.com/jagguli/assistantmanager"
"Bug Tracker" = "https://github.com/jagguli/assistantmanager/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
astmgr = "astmgr.shell:main"
